Xplorer booklet support (kinda ugly)

+++ src/builder/UserBooklets.ts
@@ -36,7 +36,7 @@ class UserBookletsStore extends GeneralizedUserItem {
             // There's no excellent place to do this, so I'm doing this here: drop WIP booklets that we no longer own.
             for (const setId in setsManager.setsInfo) {
                 const booklet_id = setsManager.setsInfo[setId].booklet;
-                if (booklet_id && this.booklets.indexOf(booklet_id) === -1)
+                if (booklet_id && !booklet_id.startsWith('tutorial/') && this.booklets.indexOf(booklet_id) === -1)
                     setsManager.deleteLocalSet(setId);
             }
         } catch(ex) {
